==> research/notebook/0_Dataset.ipynb <==
# Re-organize dataset

# import
import os
import sys
from module.config import PROJ_DIR

# func : makeDir()
def makekDir(PROJ_DIR, dir_name) :
    dir_abs_name = os.path.join(PROJ_DIR, dir_name)

    if not os.path.exists(dir_abs_name):
        os.mkdir(dir_abs_name)
    else : 
        print(f"{os.path.basename(dir_abs_name)} exists!") 

# dir 
user_dir = "data/temp/users"
restaurant_dir = "data/temp/restaurants"
rating_dir = "data/temp/ratings"

# mkdir
makekDir(PROJ_DIR, user_dir)
makekDir(PROJ_DIR, restaurant_dir)
makekDir(PROJ_DIR, rating_dir)


# func : moveFile()
def moveFile(PROJ_DIR, files) :
    from_abs_file = os.path.join(PROJ_DIR, files[0])
    to_abs_file = os.path.join(PROJ_DIR, files[1])

    if not os.path.exists(to_abs_file) :
        os.replace(from_abs_file, to_abs_file)
    else : 
        print(f"{os.path.basename(to_abs_file)} exists!") 

# define 
rating_file = [
    ("data/raw/RCdata/rating_final.csv", "data/temp/ratings/rating_final.csv")
]

user_files = [
    ("data/raw/RCdata/usercuisine.csv", "data/temp/users/usercuisine.csv"), 
    ("data/raw/RCdata/userpayment.csv", "data/temp/users/userpayment.csv"), 
    ("data/raw/RCdata/userprofile.csv", "data/temp/users/userprofile.csv")
]

restaurant_files = [
    ("data/raw/RCdata/chefmozaccepts.csv", "data/temp/restaurants/chefmozaccepts.csv"), 
    ("data/raw/RCdata/chefmozcuisine.csv", "data/temp/restaurants/chefmozcuisine.csv"), 
    ("data/raw/RCdata/chefmozhours4.csv", "data/temp/restaurants/chefmozhours4.csv"),
    ("data/raw/RCdata/chefmozparking.csv", "data/temp/restaurants/chefmozparking.csv"), 
    ("data/raw/RCdata/geoplaces2.csv", "data/temp/restaurants/geoplaces2.csv")
]

# move (no need for rating : we'll construct it)
for file in user_files + restaurant_files + rating_file : 
    moveFile(PROJ_DIR, file)
