==> research/notebook/1_FeatureEngineering.ipynb <==
# Explore Dataset

# import
import os
import pandas as pd
from functools import reduce
from module.config import PROJ_DIR
from IPython.display import display

# utils functions

## Data exploration

## Data mining
The goal is to reorganize our dataset. We analyse an element in dataset according to permanent or non permanent.

For example latitude, longitude of user will go to criterai until users can ask for many recommandations at different places. But restaurant address is fixed and will go directly to theirs. 

- **User**
- **Criteria**
- **Restaurant**

### Users

# user cuisine
user_cuisine_file = os.path.join(PROJ_DIR, "data/temp/users/usercuisine.csv")
user_cuisine_df = pd.read_csv(user_cuisine_file)

# user payment
user_payment_file = os.path.join(PROJ_DIR, "data/temp/users/userpayment.csv")
user_payment_df = pd.read_csv(user_payment_file)

# user profile
user_profile_file = os.path.join(PROJ_DIR, "data/temp/users/userprofile.csv")
user_profile_df = pd.read_csv(user_profile_file)

# merge user_df
user_df = user_profile_df.merge(user_payment_df, on='userID')
user_df = user_df.merge(user_cuisine_df, on='userID')

display(user_profile_df.columns)
display(user_payment_df.columns)
display(user_cuisine_df.columns)

display(user_df.sample(2))

# define cols
profile_cols = ['userID', 'birth_year', 'personality', 'interest', 'religion', 
                'marital_status', 'activity', 'ambience', 
                'smoker', 'drink_level', 'hijos',
                'color', 'weight', 'height'    
]

criteria_cols = ['userID', 'latitude', 'longitude', 'Upayment', 'Rcuisine',
                'budget', 'dress_preference', 'transport'
]

# dataframe
profile_df = user_df[profile_cols].copy()
criteria_df = user_df[criteria_cols].copy()

display(profile_df.sample(2))
display(criteria_df.sample(2))

### Restaurants

# restaurant accepts
resto_accepts_file = os.path.join(PROJ_DIR, "data/temp/restaurants/chefmozaccepts.csv")
resto_accepts_df = pd.read_csv(resto_accepts_file)

# restaurant cuisine
resto_cuisine_file = os.path.join(PROJ_DIR, "data/temp/restaurants/chefmozcuisine.csv")
resto_cuisine_df = pd.read_csv(resto_cuisine_file)

# restaurant hours
resto_hours_file = os.path.join(PROJ_DIR, "data/temp/restaurants/chefmozhours4.csv")
resto_hours_df = pd.read_csv(resto_hours_file)

# restaurant parking
resto_parking_file = os.path.join(PROJ_DIR, "data/temp/restaurants/chefmozparking.csv")
resto_parking_df = pd.read_csv(resto_parking_file)

# restaurant geoplace
resto_geoplace_file = os.path.join(PROJ_DIR, "data/temp/restaurants/geoplaces2.csv")
resto_geoplace_df = pd.read_csv(resto_geoplace_file , encoding='latin-1')

# merge resto_df
dfs = [resto_accepts_df, resto_cuisine_df, resto_hours_df, 
        resto_parking_df, resto_geoplace_df]
resto_df = reduce(lambda left,right: pd.merge(left,right,on='placeID'), dfs)

display(resto_df.columns)

# order columns
resto_cols = ['placeID', 'name', 'address', 'latitude', 'longitude', 
            'city', 'state', 'country', 'fax', 'zip', 'the_geom_meter', 
            'hours', 'days', 'Rpayment', 'Rcuisine', 'parking_lot',
            'alcohol', 'smoking_area', 'dress_code', 'accessibility', 
            'price','Rambience', 'franchise','area', 'other_services', 
            'url'
]

# dataframe
restaurant_df = resto_df[resto_cols].copy()
display(restaurant_df.sample(2))

### Ratings

# restaurant accepts
ratings_file = os.path.join(PROJ_DIR, "data/temp/ratings/rating_final.csv")
ratings_df = pd.read_csv(ratings_file)

display(ratings_df.sample(2))

### Export

# profile & restaurant must be unique
profile_df.drop_duplicates(subset=['userID'])

# export dataset
profile_df.to_csv(os.path.join(PROJ_DIR, "data/temp/clean/user_profile.csv"))
criteria_df.to_csv(os.path.join(PROJ_DIR, "data/temp/clean/user_criteria.csv"))
restaurant_df.to_csv(os.path.join(PROJ_DIR, "data/temp/clean/restaurant.csv"))
ratings_df.to_csv(os.path.join(PROJ_DIR, "data/temp/clean/rating.csv"))
