==> research/notebook/2_DataIngestion.ipynb <==
# Data Ingestion
Process : 
- init sqlalchemy database
- create pythonic table
- add records in table
- push in AWS RDS (later)

# import 
import os
import pandas as pd
from datetime import datetime
from module.config import PROJ_DIR
from sqlalchemy import create_engine
from sqlalchemy import Column, Integer, String, DateTime, Float
from sqlalchemy.orm import declarative_base, sessionmaker
from IPython.display import display

# dataset
profile_df = pd.read_csv(os.path.join(PROJ_DIR, "data/temp/clean/user_profile.csv"), index_col=[0])
restaurant_df = pd.read_csv(os.path.join(PROJ_DIR, "data/temp/clean/restaurant.csv"), index_col=[0])

display(profile_df.columns)
display(restaurant_df.columns)

connection_string = "sqlite:///" + os.path.join(PROJ_DIR, 'data/db/foodapp.db')

Base = declarative_base()
engine = create_engine(connection_string, echo=True, future=True)
Session = sessionmaker()


"""
class User
    userID int
    birth_year int
    personality str
    interest str
    religion str
    marital_status str
    activity str
    ambience str
    smoker str
    drink_level str
    hijos str
    color str
    weight float
    height float
"""

class User(Base) :
    __tablename__ = "users"
    userID = Column(Integer(), primary_key=True)
    # email = Column(String(100), unique=True)
    # password = Column(String(100))
    # username = Column(String(1000))

    birth_year = Column(Integer())
    personality = Column(String(100))
    interest = Column(String(100))
    religion = Column(String(100))
    marital_status = Column(String(100))
    activity = Column(String(100))
    ambience = Column(String(100))
    smoker = Column(String(100))
    drink_level = Column(String(100))
    hijos = Column(String(100))
    color = Column(String(100))
    weight = Column(Float())
    height = Column(Float())


    def __repr__(self) -> str:
        repr = f"<User ID={self.userID} color={self.color}>"
        return repr

mae = User(userID=1, weight=0.21)
print(mae)

"""
class Place
    id int
    name str
    address str
    latitude float
    longitude float
    city str
    state str
    country str
    fax int
    zip str
    the_geom_meter str
    hours str
    days str
    Rpayment str
    Rcuisine str
    parking_lot str
    alcohol str
    smoking_area str

'placeID', 'name', 'address', 'latitude', 'longitude', 'city', 'state',
       'country', 'fax', 'zip', 'the_geom_meter', 'hours', 'days', 'Rpayment',
       'Rcuisine', 'parking_lot', 'alcohol', 'smoking_area', 'dress_code',
       'accessibility', 'price', 'Rambience', 'franchise', 'area',
       'other_services', 'url'
"""

class User(Base) :
    __tablename__ = "places"
    placeID = Column(Integer(), primary_key=True)
    # email = Column(String(100), unique=True)
    # password = Column(String(100))
    # username = Column(String(1000))
