==> research/notebook/3_TrainModel.ipynb <==
# Modelling
Models :
- **DBSCAN** : clustering `restaurants` by location (lat, lon);
- **RF** : train model to predict foot rating and service rating from `profile`, `criteria`, `restaurant` & `rating`. 

## Clustering `Restaurant`

## Modelling `Review`
